Break up into three modules for plotting, staterealtions, and model
